Recolocado record Direccion
