+++ src/negocio/SAPedidosImp.java
@@ -2,6 +2,7 @@ package negocio;
 
 import integracion.DAOPedidos;
 import integracion.DAOPedidosImp;
+import utils.Direccion;
 
 import java.text.*;
 import java.util.Collection;
@@ -28,7 +29,7 @@ public class SAPedidosImp implements SAPedidos {
     private BOPedido transform(TOPedido pedido){
         BOPedido boPedido = new BOPedido();
         boPedido.setID(pedido.getID());
-        boPedido.setDireccion(BOPedido.Direccion.parse(pedido.getDireccion()));
+        boPedido.setDireccion(Direccion.parse(pedido.getDireccion()));
         boPedido.setIdCesta(pedido.getIDCesta());
         boPedido.setIdUsuario(pedido.getIDUsuario());
         boPedido.setStatus(BOPedido.Status.valueOf(pedido.getStatus()));
